requirements.txt
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.1.1